// src/generated/client/types/PostWorkoutsRequestBody.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { PostWorkoutsRequestExercise } from "./PostWorkoutsRequestExercise.ts";

export type PostWorkoutsRequestBody = {
    /**
     * @type object | undefined
    */
    workout?: {
        /**
         * @description The title of the workout.
         * @type string | undefined
        */
        title?: string;
        /**
         * @description A description for the workout workout.
         * @type string
        */
        description?: string | null;
        /**
         * @description The time the workout started.
         * @type string | undefined
        */
        start_time?: string;
        /**
         * @description The time the workout ended.
         * @type string | undefined
        */
        end_time?: string;
        /**
         * @description A boolean indicating if the workout is private.
         * @type boolean | undefined
        */
        is_private?: boolean;
        /**
         * @type array | undefined
        */
        exercises?: PostWorkoutsRequestExercise[];
    };
};