// src/generated/client/types/PostWorkoutsRequestExercise.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { PostWorkoutsRequestSet } from "./PostWorkoutsRequestSet.ts";

export type PostWorkoutsRequestExercise = {
  /**
   * @description The ID of the exercise template.
   * @type string | undefined
   */
  exercise_template_id?: string;
  /**
   * @description The ID of the superset.
   * @type integer
   */
  superset_id?: number | null;
  /**
   * @description Additional notes for the exercise.
   * @type string
   */
  notes?: string | null;
  /**
   * @type array | undefined
   */
  sets?: PostWorkoutsRequestSet[];
};